Add quick view for issue in taskboard
