+++ app/components/app-ui/task-board.hbs
@@ -47,6 +47,7 @@
             @statusClass={{@statusClass}}
             @query={{this.query}}
             @openIssue={{@openIssue}}
+            @selectIssue={{this.selectIssue}}
           />
         </div>
       {{/each}}
