+++ app/components/app-ui/task-board/milestones.hbs
@@ -40,9 +40,9 @@
             <AppUi::TaskBoard::IssuesProvider @issues={{this.milestone.issues}} @query={{@query}} as |issues|>
               {{#each issues as |issue|}}
                 {{#if (eq issue.status status.name)}}
-                  <div class="item" style="{{laneitem-color issue.issuetype.name}}" data-field-issue-id="{{issue.id}}" data-field-issue-status="{{issue.status}}" data-field-issue-milestone="{{issue.milestoneId}}">
+                  <div class="board-issue-item" style="{{laneitem-color issue.issuetype.name}}" data-field-issue-id="{{issue.id}}" data-field-issue-status="{{issue.status}}" data-field-issue-milestone="{{issue.milestoneId}}">
                   
-                    <h4>{{priority-icon priority=(readonly issue.priority)}} <LinkTo @route="app.project.issue.page" @model={{issue.issueNumber}} {{action @openIssue issue}}>#{{issue.issueNumber}} - {{issue.subject}}</LinkTo></h4>
+                    <h4>{{priority-icon priority=(readonly issue.priority)}} <a href="#" {{on "click" (fn @selectIssue issue)}}>#{{issue.issueNumber}} - {{issue.subject}}</a></h4>
                     <p class="description">{{safe-html issue.description}}</p>
                     <div class="flex-row flex-jc-end">
                       <div><img class="img-circle" src="{{this.apiHost}}/userimage/get/{{issue.assignedTo.id}}" alt="{{issue.assignedTo.name}}" title="{{issue.assignedTo.name}}"></div>
