+++ app/templates/app/project/board.hbs
@@ -1,26 +1,70 @@
 <section class="content">
   <div class="board row">
+    <div class="col-md-{{if this.selectedIssue '8' '12'}}">
+      <AppUi::TaskBoard
+        @milestones={{this.milestoneList}}
+        @statuses={{this.statuses}}
+        @statusClass={{this.statusClass}}
+        @backlog={{this.backlog}}
+        @group="milestones"
+        @scroll={{true}}
+        @scrollSensitivity={{200}}
+        @scrollSpeed = {{10}}
+        @sort={{true}}
+        @forceFallback={{true}}
+        @disableSortable={{false}}
+        @dragClass="curr-item"
+        @chosenClass="curr-item"
+        @updateIssue={{this.updateIssue}}
+        @openIssue={{this.openIssue}}
+        @selectIssue={{this.selectIssue}}
+        @savedSearches={{this.savedSearches}}
+        @applySearch={{this.applySearch}}
+        @paramsQuery={{this.query}}
+      />
+    </div>
 
-    <AppUi::TaskBoard
-      @milestones={{this.milestoneList}}
-      @statuses={{this.statuses}}
-      @statusClass={{this.statusClass}}
-      @backlog={{this.backlog}}
-      @group="milestones"
-      @scroll={{true}}
-      @scrollSensitivity={{200}}
-      @scrollSpeed = {{10}}
-      @sort={{true}}
-      @forceFallback={{true}}
-      @disableSortable={{false}}
-      @dragClass="curr-item"
-      @chosenClass="curr-item"
-      @updateIssue={{this.updateIssue}}
-      @openIssue={{this.openIssue}}
-      @savedSearches={{this.savedSearches}}
-      @applySearch={{this.applySearch}}
-      @paramsQuery={{this.query}}
-    />
-
+    {{#if this.selectedIssue}}
+      <div class="col-md-4 issue-details-container">
+        {{#if this.isLoadingIssueDetails}}
+          <div class="box box-primary">
+            <div class="box-header with-border">
+              <h3 class="box-title">{{t "views.app.board.issue.loading"}}</h3>
+            </div>
+            <div class="box-body flex-row flex-jc-c">
+              <div class="dot-pulse"></div>
+            </div>
+          </div>
+        {{else if this.selectedIssueDetails}}
+          <div class="box box-primary">
+            <div class="box-header with-border">
+              <h3 class="box-title">
+                <LinkTo @route="app.project.issue.page" @model={{this.selectedIssueDetails.issueNumber}} target="_blank">
+                  #{{this.selectedIssueDetails.issueNumber}} - {{this.selectedIssueDetails.subject}}
+                  <i class="fa fa-external-link" aria-hidden="true"></i>
+                </LinkTo>
+              </h3>
+              <div class="box-tools pull-right">
+                <button type="button" class="btn btn-box-tool" {{action this.closeIssueDetails}}>
+                  <i class="fa fa-times"></i>
+                </button>
+              </div>
+            </div>
+            <div class="box-body">
+              <Issue::IssueDetails 
+                @issue={{this.selectedIssueDetails}}
+                @activities={{null}}
+                @selectStatic={{this.selectStatic}}
+                @projectController={{this}}
+                @issueTypes={{this.issueTypes}}
+                @s_id={{null}}
+                @scrollAndHighlight={{this.scrollAndHighlight}}
+                @quickView={{true}}
+              />
+            </div>
+          </div>
+        {{/if}}
+      </div>
+    {{/if}}
   </div>
 </section>
