+++ tests/acceptance/steps/taskboard/render-milestone-steps.js
@@ -1,5 +1,5 @@
 import steps from '../steps';
-
+import { click } from '@ember/test-helpers';
 export default function (assert) {
 	return steps(assert)
 	.given('Each issue has $hours hours and $minutes minutes of $context time', function (hours, minutes, context) {
@@ -15,6 +15,11 @@ export default function (assert) {
 			});
 		});
 	})
+	.when('User clicks on first issue', async function () {
+		let issue = server.schema.issues.all().models[0];
+		let issueEl = document.querySelector(`[data-field-issue-id="${issue.id}"] h4 a`);
+		await click(issueEl);
+	})
 	.then('User should see $hours hours and $minutes minutes of $context time of first milestone', function (hours, minutes, context) {
 		let milestone = server.schema.milestones.all().models[0];
 		let milestoneEl = document.querySelector(`[data-milestone-details="${milestone.id}"]`);
@@ -22,6 +27,11 @@ export default function (assert) {
 		let contextText = context === 'spent' ? 'Spent' : 'Estimated';
 		assert.equal(timelogEl.innerText, `${contextText} ${hours} h ${minutes} m`);
 
+	})
+	.then('User should see issue details section with issue subject', function () {
+		let issue = server.schema.issues.all().models[0];
+		let issueDetails = document.querySelector('.issue-details-container .issue-details .issue-subject');
+		assert.equal(issueDetails.innerText, issue.subject);
 	})
 	.then(
 		'User should see $issuesCount issues in $status status',
